Confidence Intervals for Predictions of a GLM

Provides confidence intervals for predictions of a GLM.

PredictCI(mod, newdata, conf.level = 0.95)

Arguments

mod

the binomial model

newdata

the data to be predicted

conf.level

confidence level of the interval. Default is 0.95.

Details

The confidence intervals for predictions are calculated with the se of the model and the normal quantile.

Value

a matrix with 3 columns for the fit, the lower confidence interval and the upper confidence interval

References

https://stackoverflow.com/questions/14423325/confidence-intervals-for-predictions-from-logistic-regression

Author

Andri Signorell <andri@signorell.net>

See also

Examples

r.logit <- FitMod(diabetes ~ age, d.pima, fitfn = "logit")
head(PredictCI(r.logit, newdata=d.pima))
#>        prob       lci       uci
#> 1 0.5134500 0.4500891 0.5763816
#> 2 0.3219794 0.2883900 0.3575155
#> 3 0.3312219 0.2977057 0.3665421
#> 4 0.2377681 0.1986836 0.2818370
#> 5 0.3405965 0.3069729 0.3759054
#> 6 0.3128741 0.2790536 0.3488105