anscresid(modtools)	R Documentation

Anscombe's Residuals

Description

The fonction provides Anscombe's residuals associated with an object 'glm'.

Usage

anscresid(object, ...)

Arguments

object 	 Object of class inheriting from '"glm"'

... 	 further arguments passed to or from other methods

Details

The formulas to compute the Anscombe'sresiduals are defined as follow:is defined as follow: for gaussian family:

r_{ans} = y - μ

for inverse.gaussian family:

r_{ans} = (log(y) - log(μ))/(μ^{0.5})

for binomial family:

r_{ans} = sqrt(m) * (b(y) - b(μ)) * (μ * (1-μ))^{-1/6}

b(x) = int_0^x{(x^{-1/3} * (1 - x)^{-1/3}}

for poisson family:

r_{ans} = (3/2) * ((y^{2/3}) * μ^{-1/6} - μ^{0.5})

for Gamma family:

r_{ans} = 3 * ((y/μ)^{1/3} - 1)

Value

The function returns a numerical vector which contains the values of anscombe's residuals for each observation.

References

McCullagh P. and Nelder, J. A. (1989) Generalized Linear Models. London: Chapman and Hall.
Pierce, D. A. and Schafer, D. W. (1986) Residuals in Generalized Linear Models, Journal of the American Statistical Association, 81,396,977-986.
Abscombe 1953

See Also

glm,residuals.glm

Examples

## binomial

## poisson

## Gamma

[Package modtools version 1.0-7 Index]