bootcoef(modtools)	R Documentation

~~function to do ... ~~

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage

bootcoef(object,data,R=99,...)

Arguments

object 	 ~~Describe object here~~ attention object et different

data 	 ~~Describe data here~~

R 	 ~~Describe R here~~

... 	 ~~Describe ... here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

References

Davison, A.C. and Hinkley, D.V. (1997) Bootstrap Methods and Their Application. Cambridge University Press.
Efron et al.
Efron et al.

See Also

~~objects to See Also as help, ~~~

Examples

# add an example

[Package modtools version 1.0-7 Index]