bootvalid(modtools)	R Documentation

Validation procedure based on bootstrap for object from 'glm' or 'lm'.

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage

bootvalid(object, ...)
## Default S3 method:
bootvalid(object, data, cost = costMSE,
R = 99,method = "raw",...)
## S3 method for class 'bootcorrected':
summary(object,display=TRUE,...)

Arguments

object 	 Object of class inheriting from '"glm"'

data 	 ~~Describe data here~~

cost 	 ~~Describe cost here~~

R 	 ~~Describe R here~~

method 	 'raw' or 'corrected'

display 	 a logical value (by default, display=TRUE)

... 	 further arguments passed to or from other methods

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

References

Davison, A.C. and Hinkley, D.V. (1997) Bootstrap Methods and Their Application. Cambridge University Press.
Efron et al.
Efron et al.
Harrel 2001

See Also

~~objects to See Also as help, ~~~

Examples

## glm1

## bootvalid

## histogram of results

[Package modtools version 1.0-7 Index]