ckappa(modtools)	R Documentation

Kappa's index

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage

ckappa(x)

Arguments

x 	 a matrix

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Note

~~further notes~~

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--    or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x) 
{
    if (ncol(x) != nrow(x)) 
        stop("non covenient dimension !")
    N <- sum(x)
    sxii <- sum(diag(x))
    sxip <- apply(x, 2, sum)
    sxpi <- apply(x, 1, sum)
    k <- (N * sxii - sum(sxip * sxpi))/(N * N - sum(sxip * sxpi))
    return(k)
  }

[Package modtools version 1.0-7 Index]