cost(modtools)	R Documentation

Cost functions

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage

costAVER(y, yhat = 0)
costMAE(y, yhat = 0)
costMSE(y, yhat = 0)
costRMSE(y, yhat = 0)
costSlope(y, yhat)
costOri(y, yhat)
costR2(y, yhat)
costBIN(y, mu = 0, cutoff = 0.5)
costKappa(y, yhat, cutoff = 0.5)
costGoodCl(y,yhat, cutoff = 0.5)

Arguments

y 	 a numerical vector corresponding to the obsevred values.

yhat 	 a numerical vector corresponding to the expected values.

mu 	 a numerical vector corresponding to the expected values.

cutoff 	 a numerical value in the range [0,1] corresponding to the threshold to transform the values 'mu' into binary vector.

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Note

~~further notes~~

add descriptions and formula

References

Davidson and Hinkley, Saporta 2006

See Also

SRM,bootvalid,cv.glm

Examples

x <- rnorm(20,2,5)
y <- -6+x*3+rnorm(20)
lm1 <- lm(y~x)
costRMSE(y,lm1$fitted)

[Package modtools version 1.0-7 Index]