Diabetes survey on Pima Indians

The National Institute of Diabetes and Digestive and Kidney Diseases conducted a study on 768 adult female Pima Indians living near Phoenix.

data(d.pima)
data(d.pima2)

Format

The dataset contains the following variables

pregnant

Number of times pregnant

glucose

Plasma glucose concentration at 2 hours in an oral glucose tolerance test

diastolic

Diastolic blood pressure (mm Hg)

triceps

Triceps skin fold thickness (mm)

insulin

2-Hour serum insulin (mu U/ml)

bmi

Body mass index (weight in kg/(height in metres squared))

diabetes

Diabetes pedigree function

age

Age (years)

test

test whether the patient shows signs of diabetes (coded 0 if negative, 1 if positive)

Details

d.pima2 is the same dataset as d.pima with the only change, that invalid 0-values are replaced by NAs.

Note

This dataset has been borrowed from Julian Faraway's package:
faraway: Functions and datasets for books by Julian Faraway, 2015

Source

The data may be obtained from the package MASS.