intervals(modtools)	R Documentation

confidence and prediction/tolerance intervals for glm

Description

This method gives confidence and prediction/tolerance intervals for the expected values from a generalized linear model (object of class 'glm').

Usage

intervals(object, ...)
## S3 method for class 'glm':
intervals(object, newdata, type = "response", interval = "confidence",
method = 1, level = 0.05, ...)

Arguments

object 	 Object of class inheriting from '"glm"'

newdata 	 An optional data frame in which to look for variables with which to predict. If omitted, the data values are used.

type 	 the type of prediction required (by defaut type="response")

interval 	 Type of interval calculation.

method 	 a numerical values (by default method=1). The option 'method = 1' gives intervals based on the carry-over of the extreme values. The option 'method = 2' provides "direct" interval (see the section note for more details.

level 	 Tolerance/confidence level

... 	 further arguments passed to or from other methods

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Note

~~further notes~~ Several procedure can provide confidence (or prediction/tolerance) intervals. In the function intervals.glm, we propose the two following procedures: method 1: confidence and prediction intervals based on the carry-over of the extreme values. This method is an extrapolation of the results obtained in the linear model. for confidence intervals

sigma = sqrt{x^tVCOVx}

where for prediction intervals

sigma = sqrt{1+x^tVCOVx}

where method 2: "direct" confidence intervals

hat{y} +- epsilon_{alpha}var(hat{y})sqrt{x^tVCOVx}

where var(hat{y})=psi var()

References

~put references to the literature/web site here ~

See Also

glm,predict.glm

Examples

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--    or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (object, ...) 
{
    UseMethod("intervals")
  }

[Package modtools version 1.0-7 Index]