modtools-package(modtools)	R Documentation

Additional tools for model diagnostic and selection

Description

This package contains exhaustive regression functions and some tools to evaluate the goodness of fit and the model quality.

Details

Package: 	 modtools

Type: 	 Package

Version: 	 1.0-7

Date: 	 2009-01-14

License: 	 GPL version 2 or newer

LazyLoad: 	 yes

Depends: 	 R (>= 2.9.0),boot

Suggests: 	 car

# nothing for the moment

Author(s)

Pierre Bady <pierre.bady@free.fr>
Maintainer: Pierre Bady <pierre.bady@free.fr>

References

Saporta (2006)
Efron et al
Hinkley and Davidson (1998)
etc
to be completed

See Also

car, boot

Examples

# nothing for the moment

[Package modtools version 1.0-7 Index]