pseudoR2(modtools)	R Documentation

Pseudo-R2 for object 'glm'

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage

pseudoR2(mod0, mod, ...)
## S3 method for class 'glm':
pseudoR2(mod0, mod, option = "phi", ...)

Arguments

mod0 	 ~~Describe mod0 here~~

mod 	 ~~Describe mod here~~

option 	 ~~Describe option here~~

... 	 ~~Describe ... here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Note

~~further notes~~

References

Mcfadden 1973, estrella 1998

See Also

~~objects to See Also as help, ~~~

Examples

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--    or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (mod0, mod, ...) 
{
    UseMethod("pseudoR2")
  }

[Package modtools version 1.0-7 Index]