roc(modtools)	R Documentation

ROC functions

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage

prep.roc(obs,pred,nbval=20,method="max",subset,...)
## S3 method for class 'roc':
plot(x, type = "curve", sub,posi = c(0.8, 0.2),...)
## S3 method for class 'roc':
print(x, ...)
## S3 method for class 'roc':
summary(object, rnd = 3, type = "curve", display = TRUE, ...)

Arguments

obs 	 ~~Describe obs here~~

pred 	 ~~Describe pred here~~

nbval 	 ~~Describe nbval here~~

subset 	 ~~Describe subset here~~

x 	 ~~Describe x here~~

method 	 ~~Describe method here~~

sub 	 ~~Describe sub here~~

object 	 ~~Describe object here~~

type 	 ~~Describe method here~~

posi 	 information position

rnd 	 ~~Describe rnd here~~

display 	 ~~Describe display here~~

... 	 ~~Describe ... here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Note

With the option 'estim', the function prep.roc used the function OptimCut to define the optimal cut-off. this one can be based on several criteria:
fgoodclassif:
fkappa
fSpecSens
fSpecSens2

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples

x <- rnorm( 100 )
z <- rnorm( 100 )
w <- rnorm( 100 )
tigol <- function( x ) 1 - ( 1 + exp( x ) )^(-1)
y <- rbinom( 100, 1, tigol( 0.3 + 3*x + 5*z + 7*w ) )
ROC( form = y ~ x + z, plot="ROC" )
glm1 <- glm(y ~ x + z,family=binomial)
roc1 <- prep.roc(glm1$y,glm1$fitted)
plot(roc1)

[Package modtools version 1.0-7 Index]