scorevalid(modtools)	R Documentation

Performance curves

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage

scorevalid(y, score, ...)
## Default S3 method:
scorevalid(y, score, recal = FALSE,
  qth=seq(0, 1, length = 10),tol=1e-04,...)
## S3 method for class 'scorevalid':
print(x,digits = getOption("digits"),...)
## S3 method for class 'scorevalid':
plot(x,mgraph = NULL,...)

Arguments

y 	 ~~Describe y here~~

score 	 ~~Describe score here~~

recal 	 ~~Describe recal here~~

x 	 ~~Describe x here~~

tol 	 1e-04

qth 	 seq(0, 1, length = 10)

mgraph 	 graphic organisation

... 	 ~~Describe ... here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Note

The functions is based on R code of the excellent document proposed by Pierre-André Cornillon (see the section 'references'). for more details, you can consult the following links:
http://www.uhb.fr/sc_sociales/labstats/PAC/doc/score.pdf
http://www.uhb.fr/sc_sociales/labstats/PAC/

References

bardos (2001)
Cornillon P-A (200x) Discrimination et Scores, MASS course - Rennes 2:
http://www.uhb.fr/sc_sociales/labstats/PAC/doc/score.pdf

See Also

roc

Examples

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--    or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (y, score, ...) 
{
    UseMethod("scorevalid")
  }

[Package modtools version 1.0-7 Index]