sefit(modtools)	R Documentation

Standard error of prediction (or fitted values) from glm

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage

sefit(object, ...)
## S3 method for class 'glm':
sefit(object, newdata, interval = "confidence", dispersion = TRUE,m=1,...)

Arguments

object 	 Object of class inheriting from '"glm"'

newdata 	 An optional data frame in which to look for variables with which to predict. If omitted, the data values are used.

interval 	 a logical value ...

dispersion 	 Type of interval calculation.

m 	 a numerical value corresponding to observation number

... 	 further arguments passed to or from other methods

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Author(s)

~~who you are~~

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--    or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (object, ...) 
{
    UseMethod("sefit")
  }

[Package modtools version 1.0-7 Index]