Control Parameters for Zero-inflated Count Data Regression

Various parameters that control fitting of zero-inflated regression models using zeroinfl.

zeroinfl.control(method = "BFGS", maxit = 10000, trace = FALSE,
  EM = FALSE, start = NULL, ...)

Arguments

method

characters string specifying the method argument passed to optim.

maxit

integer specifying the maxit argument (maximal number of iterations) passed to optim.

trace

logical or integer controlling whether tracing information on the progress of the optimization should be produced (passed to optim).

EM

logical. Should starting values be estimated by the EM (expectation maximization) algorithm? See details.

start

an optional list with elements "count" and "zero" (and potentially "theta") containing the coefficients for the corresponding component.

...

arguments passed to optim.

Details

All parameters in zeroinfl are estimated by maximum likelihood using optim with control options set in zeroinfl.control. Most arguments are passed on directly to optim, only trace is also used within zeroinfl and EM/start control the choice of starting values for calling optim.

Starting values can be supplied, estimated by the EM (expectation maximization) algorithm, or by glm.fit (the default). Standard errors are derived numerically using the Hessian matrix returned by optim. To supply starting values, start should be a list with elements "count" and "zero" and potentially "theta" (for negative binomial components only) containing the starting values for the coefficients of the corresponding component of the model.

Value

A list with the arguments specified.

Author

Achim Zeileis <Achim.Zeileis@R-project.org>

See also

Examples

if (FALSE) {
data("bioChemists", package = "pscl")

## default start values
fm1 <- zeroinfl(art ~ ., data = bioChemists)

## use EM algorithm for start values
fm2 <- zeroinfl(art ~ ., data = bioChemists, EM = TRUE)

## user-supplied start values
fm3 <- zeroinfl(art ~ ., data = bioChemists,
  start = list(count = c(0.7, -0.2, 0.1, -0.2, 0, 0), zero = -1.7))
}